requirements.txt
flask
flask-cors
flask-socketio
psycopg2-binary
gunicorn
gevent
gevent-websocket
eventlet
python-dotenv
